+++ release-notes.md
@@ -1,5 +1,6 @@
 `1.9.4` - ?
 * Fix for missing HTML tag on action card
+* List cards dealt in tattle (in addition to drawn)
 
 `1.9.3` - 2023-11-02
 * Add reaction annotation for cards in combat manager ([#62](https://github.com/wrycu/paranoia2017/issues/62))